requirements.txt
# Core dependencies for webhook server
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv
pydantic==2.11.7

# Telegram bot (for sending confirmation messages)
aiogram>=3.0.0

# Supabase client
supabase>=2.0.0
postgrest>=0.10.0

# Stripe payments
stripe>=5.0.0

# HTTP requests
requests>=2.31.0
httpx>=0.27.0,<0.29.0

# Async support
aiohttp